Change PTransform to DoFn
